Final version. Grade : 125/100

+++ ft_strlcpy.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
 {
 	unsigned int	i;
 	unsigned int	sizereturn;
